+++ main.go
@@ -1,25 +1,8 @@
 package main
 
 import (
-	"net/http"
-
 	"github.com/gin-gonic/gin"
 )
-
-// album represents data about a record album
-type album struct {
-	ID     string  `json:"id"`
-	Title  string  `json:"title"`
-	Artist string  `json:"artist"`
-	Price  float64 `json:"price"`
-}
-
-// albums slice to seed record album data
-var albums = []album{
-	{ID: "1", Title: "Blue Train", Artist: "John Coltrane", Price: 56.99},
-	{ID: "2", Title: "Jeru", Artist: "Gerry Mulligan", Price: 17.99},
-	{ID: "3", Title: "Sarah Vaughan and Clifford Brown", Artist: "Sarah Vaughan", Price: 39.99},
-}
 
 func main() {
 	router := gin.Default()
@@ -28,38 +11,4 @@ func main() {
 	router.POST("/albums", postAlbums)
 
 	router.Run("localhost:8080")
-}
-
-// getAlbums responds with the list of all albums as JSON
-func getAlbums(c *gin.Context) {
-	c.IndentedJSON(http.StatusOK, albums)
-}
-
-// function to iterate through all albums searching for album with matching supplied id.
-func getAlbumByID(c *gin.Context) {
-	id := c.Param("id")
-
-	// loop through list of albums looking for matching id
-	for _, a := range albums {
-		if a.ID == id {
-			c.IndentedJSON(http.StatusOK, a)
-			return
-		}
-	}
-	// if no match is found return friendly album not found message
-	c.IndentedJSON(http.StatusNotFound, gin.H{"message": "album not found"})
-}
-
-// PostAlbums adds an album from JSON received in the request body
-func postAlbums(c *gin.Context) {
-	var newAlbum album
-
-	// Call BindJSON to bind the received JSON to newAlbum
-	if err := c.BindJSON(&newAlbum); err != nil {
-		return
-	}
-
-	// Add the new album to the slice
-	albums = append(albums, newAlbum)
-	c.IndentedJSON(http.StatusCreated, newAlbum)
 }
